mods to handle line-printer output converted to UTF-8
This is used to handle print-files in non-standard character sets
and create PDFs from them. I use it for retrocomputing.

New behavior: raw carriage returns now return the line to start, but
don’t progress. This allows overprinting, which used to be common.
It does not support the **bold** effect that old printers gave
for multiple impressions, but it does handle overstrikes and underlining.

--trunc option can be set to ignore any content not on the page, rather than wrapping it.